Allow mixing shipping and pickup on blocks checkout
