+++ plugins/woocommerce-blocks/assets/js/blocks/checkout/inner-blocks/checkout-pickup-options-block/style.scss
@@ -1,79 +1,77 @@
-.wc-block-checkout__pickup-options,
+.wc-block-checkout__options,
 .wp-block-woocommerce-checkout-pickup-options-block {
-	.wc-block-components-local-pickup-rates-control {
-		.wc-block-components-radio-control__option {
-			margin: 0;
-			padding: em($gap) em($gap-small) em($gap) em($gap-huge);
-		}
-		.wc-block-components-shipping-rates-control__no-results-notice {
-			margin: em($gap-small) 0;
-		}
-		.wc-block-components-radio-control
-		.wc-block-components-radio-control__input {
-			top: auto;
-			transform: none;
-		}
-		.wc-block-components-radio-control__option-layout {
-			display: block;
-		}
-		.wc-block-components-radio-control__label-group {
-			width: 100%;
-			display: flex;
+	.wc-block-components-radio-control__option {
+		margin: 0;
+		padding: .875em .875em .875em 3.5em;
+	}
+	.wc-block-components-shipping-rates-control__no-results-notice {
+		margin: em($gap-small) 0;
+	}
+	.wc-block-components-radio-control
+	.wc-block-components-radio-control__input {
+		top: auto;
+		transform: none;
+	}
+	.wc-block-components-radio-control__option-layout {
+		display: block;
+	}
+	.wc-block-components-radio-control__label-group {
+		width: 100%;
+		display: flex;
 
-			> :last-child {
-				margin-left: auto;
-			}
+		> :last-child {
+			margin-left: auto;
 		}
-		.wc-block-components-radio-control__label-group {
-			em {
-				text-transform: uppercase;
-				font-style: inherit;
-			}
+	}
+	.wc-block-components-radio-control__label-group {
+		em {
+			text-transform: uppercase;
+			font-style: inherit;
 		}
-		.wc-block-components-radio-control__description-group {
-			display: block;
-			width: calc(100% + em($gap-huge / 0.875));
-			box-sizing: border-box;
-			border-radius: $universal-border-radius;
-			padding: 0 em($gap);
-			margin-left: -(em($gap-huge));
-			margin-top: em($gap);
-			@include font-size(regular);
+	}
+	.wc-block-components-radio-control__description-group {
+		display: block;
+		width: calc(100% + em($gap-huge / 0.875));
+		box-sizing: border-box;
+		border-radius: $universal-border-radius;
+		padding: 0 em($gap);
+		margin-left: -(em($gap-huge));
+		margin-top: em($gap);
+		@include font-size(regular);
+	}
+	.wc-block-components-radio-control__description,
+	.wc-block-components-radio-control__secondary-description {
+		width: 100%;
+		text-align: left !important;
+		margin: 0;
+		display: block;
+		@include font-size(regular);
+	}
+	.wc-block-components-radio-control__description + .wc-block-components-radio-control__secondary-description {
+		margin-top: em($gap-small);
+	}
+	.wc-block-components-radio-control__description {
+		color: $gray-700;
+
+		.has-dark-controls & {
+			color: $gray-300;
 		}
-		.wc-block-components-radio-control__description,
+
+		> svg {
+			vertical-align: middle;
+			margin-top: -4px;
+			fill: currentColor;
+		}
+	}
+	.wc-block-components-radio-control__description-group {
 		.wc-block-components-radio-control__secondary-description {
-			width: 100%;
-			text-align: left;
-			margin: 0;
-			display: block;
-			@include font-size(regular);
+			display: none;
 		}
-		.wc-block-components-radio-control__description + .wc-block-components-radio-control__secondary-description {
-			margin-top: em($gap-small);
-		}
-		.wc-block-components-radio-control__description {
-			color: $gray-700;
-
-			.has-dark-controls & {
-				color: $gray-300;
-			}
-
-			> svg {
-				vertical-align: middle;
-				margin-top: -4px;
-				fill: currentColor;
-			}
-		}
+	}
+	.wc-block-components-radio-control__option-checked {
 		.wc-block-components-radio-control__description-group {
 			.wc-block-components-radio-control__secondary-description {
-				display: none;
-			}
-		}
-		.wc-block-components-radio-control__option-checked {
-			.wc-block-components-radio-control__description-group {
-				.wc-block-components-radio-control__secondary-description {
-					display: block;
-				}
+				display: block;
 			}
 		}
 	}
