+++ plugins/woocommerce-blocks/assets/js/blocks/checkout/inner-blocks/checkout-shipping-methods-block/block.tsx
@@ -12,54 +12,12 @@ import {
 	hasCollectableRate,
 	isAddressComplete,
 } from '@woocommerce/base-utils';
-import { getCurrencyFromPriceResponse } from '@woocommerce/price-format';
-import {
-	FormattedMonetaryAmount,
-	StoreNoticesContainer,
-} from '@woocommerce/blocks-components';
+import { StoreNoticesContainer } from '@woocommerce/blocks-components';
 import { useEditorContext, noticeContexts } from '@woocommerce/base-context';
-import { decodeEntities } from '@wordpress/html-entities';
-import { getSetting } from '@woocommerce/settings';
-import type {
-	PackageRateOption,
-	CartShippingPackageShippingRate,
-} from '@woocommerce/types';
 import NoticeBanner from '@woocommerce/base-components/notice-banner';
 import type { ReactElement } from 'react';
 import { useMemo } from '@wordpress/element';
-
-/**
- * Renders a shipping rate control option.
- *
- * @param {Object} option Shipping Rate.
- */
-const renderShippingRatesControlOption = (
-	option: CartShippingPackageShippingRate
-): PackageRateOption => {
-	const priceWithTaxes = getSetting( 'displayCartPricesIncludingTax', false )
-		? parseInt( option.price, 10 ) + parseInt( option.taxes, 10 )
-		: parseInt( option.price, 10 );
-
-	const secondaryLabel =
-		priceWithTaxes === 0 ? (
-			<span className="wc-block-checkout__shipping-option--free">
-				{ __( 'Free', 'woocommerce' ) }
-			</span>
-		) : (
-			<FormattedMonetaryAmount
-				currency={ getCurrencyFromPriceResponse( option ) }
-				value={ priceWithTaxes }
-			/>
-		);
-
-	return {
-		label: decodeEntities( option.name ),
-		value: option.rate_id,
-		description: decodeEntities( option.description ),
-		secondaryLabel,
-		secondaryDescription: decodeEntities( option.delivery_time ),
-	};
-};
+import renderDefaultRadioOption from '@woocommerce/base-utils/render-default-radio-option';
 
 const NoShippingAddressMessage = () => {
 	return (
@@ -149,7 +107,7 @@ const Block = ( {
 							) }
 						</>
 					}
-					renderOption={ renderShippingRatesControlOption }
+					renderOption={ renderDefaultRadioOption }
 					collapsible={ false }
 					shippingRates={ filteredShippingRates }
 					isLoadingRates={ isLoadingRates }
