+++ plugins/woocommerce-blocks/assets/js/blocks/checkout/inner-blocks/checkout-shipping-methods-block/edit.tsx
@@ -60,7 +60,7 @@ export const Edit = ( {
 			attributes={ attributes }
 			setAttributes={ setAttributes }
 			className={ clsx(
-				'wc-block-checkout__shipping-option',
+				'wc-block-checkout__options',
 				attributes?.className
 			) }
 		>
