--- plugins/woocommerce-blocks/assets/js/blocks/checkout/inner-blocks/checkout-shipping-methods-block/style.scss
@@ -1,21 +1,0 @@
-.wc-block-checkout__shipping-option {
-	.wc-block-components-radio-control__option {
-		margin: 0;
-		padding: 0.875em 0.875em 0.875em em($gap-huge);
-	}
-
-	.wc-block-checkout__shipping-option--free {
-		text-transform: uppercase;
-	}
-
-	.wc-block-components-shipping-rates-control__no-results-notice {
-		margin: em($gap-small) 0;
-	}
-
-	.wc-block-components-shipping-rates-control__no-shipping-address-message {
-		background-color: $gray-200;
-		color: $gray-700;
-		padding: em($gap-large);
-		text-align: center;
-	}
-}
