+++ plugins/woocommerce-blocks/packages/components/form-step/style.scss
@@ -100,7 +100,7 @@
 	}
 }
 
-.wc-block-checkout__shipping-option .wc-block-components-checkout-step__content,
+.wp-block-woocommerce-checkout-shipping-methods-block .wc-block-components-checkout-step__content,
 .wc-block-checkout__payment-method .wc-block-components-checkout-step__content {
 	padding-top: $gap-smaller;
 }
